replace old input with mod editor
