+++ src/components/view/view-tweet.tsx
@@ -209,8 +209,8 @@ export function ViewTweet(tweet: ViewTweetProps): JSX.Element {
         </div>
         {user?.keyPair && (
           <Input
-            reply
-            parent={{ id: tweetId, username: username, userId: ownerId }}
+            isReply
+            parentPost={{ id: tweetId, username: username, userId: ownerId }}
             parentUrl={topic?.url || undefined}
           />
         )}
